--- TO-Do-List/src/index.css
@@ -1,83 +1,0 @@
-:root {
-  font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
-  line-height: 1.5;
-  font-weight: 400;
-  color-scheme: light dark;
-  color: #333;
-  background-color: #f2f2f2;
-  text-rendering: optimizeLegibility;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-  -webkit-text-size-adjust: 100%;
-  height: 100vh;
-  display: flex;
-  align-items: flex-start; 
-  justify-content: center;
-  padding-top: 20px; 
-}
-
-a {
-  font-weight: 500;
-  color: #007bff;
-  text-decoration: none;
-}
-
-a:hover {
-  color: #0056b3;
-}
-
-body {
-  margin: 0;
-}
-
-.container {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-h1 {
-  font-size: 2.8em;
-  line-height: 1.2;
-  margin-bottom: 20px;
-}
-
-button {
-  border-radius: 12px;
-  border: 10px;
-  padding: 10px 20px;
-  font-size: 1em;
-  font-weight: 500;
-  font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
-  background-color: #007bff;
-  color: #fff;
-  cursor: pointer;
-  transition: background-color 0.25s;
-  margin-top: 10px;
-}
-
-button:hover {
-  background-color: #0056b3;
-}
-
-button:focus,
-button:focus-visible {
-  outline: none;
-}
-
-@media (prefers-color-scheme: light) {
-  :root {
-    color: #333;
-    background-color: #fff;
-  }
-
-  a:hover {
-    color: #0056b3;
-  }
-
-  button {
-    background-color: lightgreen;
-    color: #333;
-    gap: 10px;
-  }
-}
