+++ TO-Do-List/src/pages/About.jsx
@@ -1,0 +1,11 @@
+export default function About() {
+    return (
+      <div>
+        <h1 className="about-title">About</h1>
+        <p className="about">
+          This is a simple To-Do List application built using React. Users can add tasks to the list, delete tasks, and mark tasks as completed from the list.
+        </p>
+      </div>
+    );
+  }
+
